requirements.txt
# CLI Framework
click==8.1.7

# Notion API
notion-client==2.2.1

# HTTP & API
requests==2.31.0

# Data Validation
pydantic>=2.10.0

# Environment Variables
python-dotenv==1.0.0

# CLI Output & Formatting
rich==13.7.0
tabulate==0.9.0

# Date & Time
python-dateutil==2.8.2

# Development Tools
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.1
flake8==6.1.0
